+++ CHANGELOG.md
@@ -54,6 +54,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix no-relative-imports issues (ticket-4de43b5c)
 - Fix ast-unused-imports issues (ticket-bf08772a)
 
+## [0.4.197] - 2026-07-06
+
+### Docs
+- Update README.md
+
+### Other
+- Update adapters/python/urirun/host/_dashboard_get_handlers.py
+- Update adapters/python/urirun/host/work_queue.py
+
 ## [0.4.191]
 
 ### Added